chore(eslint-plugin): revert: report function head
