+++ packages/eslint-plugin/src/rules/promise-function-async.ts
@@ -91,7 +91,6 @@ export default util.createRule<Options, MessageIds>({
     ]);
     const parserServices = util.getParserServices(context);
     const checker = parserServices.program.getTypeChecker();
-    const sourceCode = context.getSourceCode();
 
     function validateNode(
       node:
@@ -141,14 +140,12 @@ export default util.createRule<Options, MessageIds>({
         return context.report({
           messageId: 'missingAsync',
           node,
-          loc: util.getFunctionHeadLoc(node, sourceCode),
         });
       }
 
       context.report({
         messageId: 'missingAsync',
         node,
-        loc: util.getFunctionHeadLoc(node, sourceCode),
         fix: fixer => {
           if (
             node.parent &&
